Add fiftyone tool script and two COCO dataset subsets
The 500 image COCO dataset subset is generated directly by the fiftyone dataset exploration tool, querying only the label dog for a maximum of 500 samples. The 207 image COCO dataset subset is a manually cleaned version of the 500 image dataset subset. The cleaning was performed by Akshay Paruchuri and targeted the removal of images with multiple dogs, too many objects of interest, or presented the object of interest in an especially challenging way (e.g., dog was a very, very small portion of the image).